refactor(Carousel): Move some files around
