+++ src/atoms/CarouselBar.js
