--- src/atoms/carousel/index.js
@@ -1,5 +1,0 @@
-export { CarouselBar } from './CarouselBar';
-export { CarouselSlot } from './CarouselSlot';
-export { CarouselSlotList } from './CarouselSlotList';
-export { CarouselButton } from './CarouselButton';
-export { CarouselSection } from './CarouselSection';
