+++ src/index.js
@@ -64,6 +64,11 @@ export { YoutubeFrame            } from './atoms/MainRecipe/youtubeFrame';
 export { Caption                 } from './atoms/Caption';
 export { Cite                    } from './atoms/Cite';
 export { Tag                     } from './atoms/Tag';
+export { CarouselBar             } from './atoms/CarouselBar';
+export { CarouselSlot            } from './atoms/CarouselSlot';
+export { CarouselSlotList        } from './atoms/CarouselSlotList';
+export { CarouselButton          } from './atoms/CarouselButton';
+export { CarouselSection         } from './atoms/CarouselSection';
 export { Col                     } from './atoms/Col';
 export { ColorBox                } from './atoms/ColorBox';
 export { DontPushTheAdBoundaries } from './atoms/DontPushTheAdBoundaries';
